feat: update pack (controller and usecases)

+++ api/src/FradminDomain/DTOs/AdminDTOs.cs
@@ -2,4 +2,4 @@ namespace FradminDomain.DTOs;
 
 public record AdminAddDto(string Username, string Email, string? Password, short GroupId, bool IsActive);
 
-public record AdminFullDto(int Id, string Username, string Email, short GroupId, bool IsActive);
+public record AdminFullDto(int Id, string Username, string Email, int GroupId, bool IsActive);
